Refactor ghutil, add pullrequest
